GP: clean up the compilation messages
When the GlobalPlatform test suite is enabled, the build produces
*tons* of warnings. The build also prints some useless messages and
is silent on many operations.

This patch improves the situation by doing the following:
- Fix all the compiler warnings I found with GCC 4.9 (32- and 64-bit
  builds, for both TAs and xtest). Unused parameters, use of size_t
  when uint32_t is expected, constness of the TEE_GetInstanceData()
  parameter, copying a (64-bit) handle into a 32-bit value
  (params[x].value.a) and possibly a few others I don't remember.
- Remove "INFO:" messages
- Use the Kbuild format ("  OPER    path/to/target/file") to trace
  most build steps
Some more cleanup is required to avoid rebuilding stuff each time,
but that can be addressed later.

+++ host/xtest/xml/include/xml_datastorage_api.h
@@ -675,7 +675,7 @@ static TEEC_UUID UUID_TTA_testingInternalAPI_dataStorage = {
 					    connectionMethod, connectionData, \
 					    operation, &ret_orig)); \
 		if ((returnOrigin != 0) && \
-		    ((int)returnOrigin != TEEC_ORIGIN_ANY_NOT_TRUSTED_APP)) \
+		    (returnOrigin != TEEC_ORIGIN_ANY_NOT_TRUSTED_APP)) \
 			ADBG_EXPECT(c, (int)returnOrigin, ret_orig); \
 		else \
 			ADBG_EXPECT_NOT(c, (int)returnOrigin, ret_orig); \
